Reingeneered game views
Before, the navigation was done through URL parameters (es. game/gameid/upload). This was kinda bad because it allowed someone to navigate improperly through the steps by manipulating the URL. Now it's done internally through a service. (probably it's still not optimal but this is what I came up with)